WIP Typescript conversion
