+++ packages/emoji-mart/src/components/Emoji/Emoji.tsx
