--- packages/emoji-mart/src/components/Emoji/index.js
@@ -1,2 +1,0 @@
-export { default as Emoji } from './Emoji.js'
-export { default as EmojiElement } from './EmojiElement.js'
