+++ packages/emoji-mart/src/components/Emoji/index.ts
@@ -1,0 +1,2 @@
+export { default as Emoji } from './Emoji'
+export { default as EmojiElement } from './EmojiElement'
