+++ packages/emoji-mart/src/components/HOCs/PureInlineComponent.ts
