--- packages/emoji-mart/src/components/HOCs/index.js
@@ -1,1 +1,0 @@
-export { default as PureInlineComponent } from './PureInlineComponent.js'
