+++ packages/emoji-mart/src/components/HOCs/index.ts
@@ -1,0 +1,1 @@
+export { default as PureInlineComponent } from './PureInlineComponent'
