--- packages/emoji-mart/src/components/HTMLElement/index.js
@@ -1,2 +1,0 @@
-export { default as HTMLElement } from './HTMLElement.js'
-export { default as ShadowElement } from './ShadowElement.js'
