+++ packages/emoji-mart/src/components/HTMLElement/index.ts
@@ -1,0 +1,2 @@
+export { default as HTMLElement } from './HTMLElement'
+export { default as ShadowElement } from './ShadowElement'
