--- packages/emoji-mart/src/components/Navigation/index.js
@@ -1,1 +1,0 @@
-export { default as Navigation } from './Navigation.js'
