+++ packages/emoji-mart/src/components/Picker/Picker.js
