--- packages/emoji-mart/src/components/Picker/index.js
@@ -1,3 +1,0 @@
-export { default as Picker } from './Picker.js'
-export { default as PickerElement } from './PickerElement.js'
-export { default as PickerStyles } from 'bundle-text:./PickerStyles.scss'
