+++ packages/emoji-mart/src/components/Picker/index.ts
@@ -1,0 +1,3 @@
+export { default as Picker } from './Picker'
+export { default as PickerElement } from './PickerElement'
+export * as PickerStyles from 'bundle-text:./PickerStyles.scss'
