+++ packages/emoji-mart/src/declaration.d.ts
@@ -1,0 +1,1 @@
+declare module '*.scss';
