+++ packages/emoji-mart/src/helpers/frequently-used.ts
@@ -18,9 +18,9 @@ const DEFAULTS = [
   'heart',
 ]
 
-let Index = null
+let Index: any | null = null
 
-function add(emoji) {
+function add(emoji: {id: string}) {
   Index || (Index = Store.get('frequently') || {})
 
   const emojiId = emoji.id || emoji
