--- packages/emoji-mart/src/helpers/index.js
@@ -1,16 +1,0 @@
-export { default as Store } from './store.js'
-
-export { default as NativeSupport } from './native-support.js'
-export { default as FrequentlyUsed } from './frequently-used.js'
-export { default as SearchIndex } from './search-index.js'
-
-export const SafeFlags = [
-  'checkered_flag',
-  'crossed_flags',
-  'pirate_flag',
-  'rainbow-flag',
-  'transgender_flag',
-  'triangular_flag_on_post',
-  'waving_black_flag',
-  'waving_white_flag',
-]
