+++ packages/emoji-mart/src/helpers/native-support.ts
