+++ packages/emoji-mart/src/helpers/search-index.ts
