+++ packages/emoji-mart/src/helpers/store.ts
@@ -1,10 +1,10 @@
-function set(key, value) {
+function set(key: string, value: string) {
   try {
     window.localStorage[`emoji-mart.${key}`] = JSON.stringify(value)
   } catch (error) {}
 }
 
-function get(key) {
+function get(key: string): any {
   try {
     const value = window.localStorage[`emoji-mart.${key}`]
