+++ packages/emoji-mart/src/icons.tsx
