+++ packages/emoji-mart/src/index.ts
