+++ packages/emoji-mart/tsconfig.json
@@ -9,9 +9,13 @@
     "jsxImportSource": "preact",
     "isolatedModules": true,
     "strict": true,
-    "skipLibCheck": true
+    "skipLibCheck": true,
+    "jsx": "react",
+    "resolveJsonModule": true
   },
   "include": [
     "src/**/*.ts",
-  ]
+    "src/**/*.scss"
+  ],
+  "exclude": ["node_modules", "dist"],
 }
